finish vc4 mmap

+++ kernel/src/drivers/gpu/gpu_device.rs
@@ -5,9 +5,9 @@ use alloc::collections::BTreeMap;
 use alloc::sync::Arc;
 
 pub struct gpu_bo {
-	pub size: usize,
+	pub size: u32,
 	pub handle: u32,
-	pub paddr: u32,
+	pub paddr: usize,
 	pub vaddr: usize,
 	pub bo_type: u32,
 }
